requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.9.0.80
numpy==1.26.4
pdf2image==1.16.3
layoutparser==0.3.4
scikit-image==0.22.0
